chroot: Add FreeBSD support for run with chroot isolation
This copies a large amount of code from run_linux.go. Later diffs in
this stack will factor out duplicate code where possible.

Terminal handling is implemented using the posix_openpt API. We could
use the Linux implementation which uses the /dev/ptmx but that is not
present on standard FreeBSD installs - its supplied by an optional
kernel module. Conversely, posix_openpt could be used for both platforms
but has a downside of requiring cgo so its probably better to use this
only on FreeBSD.

Signed-off-by: Doug Rabson <dfr@rabson.org>