[release-1.11-rhel] Fix ADD with URL
Podman 1.6.4 and Buildah 1.11 in RHEL 7.9 had a fixed dragged
in accidentally to handle a Containerfile ADD command that had
a URL in it as part of a separate unrelated fix.  When we
cleaned up the repo recently, that fix was removed
as part of the cleanup effort as it had no corresponding
issue or BZ.

During testing, this issue popped up.  Now that we have a BZ to
attribute it to, putting the fix back in place "legally".

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2112217

and backports:  https://github.com/containers/buildah/pull/2019 which
was spun up from this issue:  https://github.com/containers/podman/issues/4686

At the time of the fix, Buildah v1.7 was under development and the fix
was put into place there and for any newer versions.

Signed-off-by: tomsweeneyredhat <tsweeney@redhat.com>