Cirrus: Disable CI testing on Ubuntu
Unfortunately much bit-rot has infected CI on this branch.  Without
lots of work there is no hope of resurrecting Ubuntu VMs for CI.  Mainly
this is due to runtime package installs, on an EOL release.  The
repositories have long-since been removed, and I'm loathe to manually
fiddle with the images to make alternatives function.  Primarily because
future buildah releases on this distro will never happen.  "Fix" this by
simply disabling Ubuntu in CI.

Signed-off-by: Chris Evich <cevich@redhat.com>