Makefile: always call $(GO) instead of `go`
In a couple of places, we were still calling `go` directly instead of
through the `$(GO)` variable, which was a mistake.  Drop one of them
entirely because the resulting value wasn't being used anywhere.

Update the definitions of `CNI_COMMIT` and `RUNC_COMMIT` to search
modules.txt instead of go.mod, which can be overridden and which might
therefore produce incorrect values.

[NO NEW TESTS NEEDED]

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>