build: honor <Containerfile>.containerignore as ignore file
As per buildkit backend documentation when using the BuildKit backend, docker build searches for
a `.dockerignore` file relative to the Dockerfile name. For example, running `docker build -f myapp.Dockerfile .`
will first look for an ignore file named `myapp.Dockerfile.dockerignore`. If such a file is not found, the
`.dockerignore` file is used if present. Using a Dockerfile based `.dockerignore` is useful if a project contains
multiple Dockerfiles that expect to ignore different sets of files.

Following PR introduces this buildkit feature parity.

Closes: https://github.com/containers/buildah/issues/4236

Signed-off-by: Aditya R <arajan@redhat.com>