pkg/cli: improve completion test
When we have a bool flag there is zero reason to define a completion
function since this flag takes no argument and the cobra lib will thus
not use it.

Podman checks that already and recently buildah caused the test to fail. [1]
To make sure this does not happen again we should have the same check here.

[1] https://github.com/containers/podman/pull/13808#issuecomment-1250318199

Signed-off-by: Paul Holzinger <pholzing@redhat.com>