Stop using ubi8
First: because we have a bunch of tests that rely on the image
manifest list, and ubi8 is not under our control: we've already
seen where ubi8 gets changed without our knowledge, requiring
almost a week of scrambling to get things fixed again.

Second, the registry it lives on is unreliable.

Signed-off-by: Ed Santiago <santiago@redhat.com>