run,performance: check build stage status in exponential backoff
Buildah constantly polls running build container for it status every
`0.1` second, this polling nature wastes CPU cycles unnecessarily so
instead of checking status every `0.1` second use exponentially
increasing counter.

Exponential counter has a max limit of `2 seconds` and delay will reset
to certain milliseconds and will grow again to `2 seconds` this process
will be repeated this status is returned from the build container.

[NO NEW TESTS NEEDED]
[NO TESTS NEEDED]
This is not a new feature but improvement to buildah so it stops
consuming and producing unnecessary overhead for jobs with longer
duration.

Closes: https://github.com/containers/buildah/issues/4481

Signed-off-by: Aditya R <arajan@redhat.com>