Fix multi-arch manifest-list build timeouts
As of this commit, the build script runs the process on x86_64 hardware
(4cpus/8gig) using emulation for foreign architectures.  Due to various
delays, inefficiencies, and other emulation-related slowdowns, it's
possible for dnf to timeout and/or mis-measure a minimum repository
download rate.  Bump up the minimum download rate by 10x and timeout
interval by double.  This should allow the builds to complete more
consistently.

Signed-off-by: Chris Evich <cevich@redhat.com>