Cap suffixDigitsModulo to a 9-digits suffix.
Because Go integers silently overflow, and 10 is multiple of 2,
repeatedly multiplying by 2 eventually turns suffixDigitsModulo
into all-bits-zero, triggering a division by zero.

So, cap the suffix to a 9-digit one. That is very likely to be sufficient.

(But don't actually limit the number of retries. We could do that as well,
for extra robustness. OTOH that would be another bit of code that we don't
test...)

[NO NEW TESTS NEEDED]
We could plausibly unit-test the loop just to see that it doesn't crash
on 100 conflicts in a row, but that's an awfully specific test that would
require moving the loop to a separate function with a closure parameter,
a complex test that makes the code less readable for a very specific crash.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>