run: drop Pdeathsig
it is not safe to use prctl(PR_SET_PDEATHSIG) in a Go program since
the parent of the process is considered the *thread* that created the
child process, and the Go runtime can free a thread at any time.   We
would either need to lock the thread, or simpler just not use it.

[NO NEW TESTS NEEDED]

Closes: https://github.com/containers/podman/issues/17918

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>