executor: apply label to only final stage
In https://github.com/containers/buildah/pull/4673 we made a change were
we were applying labels to end of each stage, which is different than
what we were doing before i.e applying label at the end of the each
step.

However buildkit does not adds label to any stage or steps it only adds
label at the end of final stage so lets do that.

Closes: https://github.com/containers/buildah/issues/4804

Signed-off-by: Aditya R <arajan@redhat.com>