CI: Support testing w/ podman-next COPR packages
Sometimes important updates need to be made to dependent packages and
run through CI w/o waiting for package release and new CI VM image
builds.  Support this in buildah CI as in podman CI, by updating
packages during setup when the magic string is present and PR is in
draft-mode.

Note: To support containerized testing, both `CIRRUS_CHANGE_TITLE`
and `CIRRUS_PR_DRAFT` env. vars. are passed through.  For these tasks,
this will result in **TWO** updates - One for the host, and another one
in the container.

Signed-off-by: Chris Evich <cevich@redhat.com>