[release-1.24] Explicitly ref. quay images for CI
Commit d5b77bf72ca1 updated the conformance test references from a
deprecated repository.  However, by pointing at the docker-hub it
inadvertantly introduce a significant possibility for flakes.  This is
because anonymous docker hub access is rate-limited by IP.  We cannot
predict the IP used for CI VMs & Containers, any of which could be at or
close to the limit.  Fix this by pointing conformance tests explicitly at
`quay.io/libpod/centos` which is excluesively for use by CI, with
static images.

Signed-off-by: Chris Evich <cevich@redhat.com>