[CI:BUILD] Packit: fix pre-sync action for downstream tasks
One of the reasons the last propose-downstream task failed for Fedora
was the `golist` tool wasn't available in the Packit environment.

This commit adds golist to the environment by downloading and extracting
the golist rpm.

This dependency could've been added in packit's upstream config but
there were a few blockers, so it's easiest to add them in our action
script.

Ref: https://github.com/containers/buildah/issues/4904

Also make go-md2man dependency in rpm/buildah.spec more generic

[NO NEW TESTS NEEDED]

Signed-off-by: Lokesh Mandvekar <lsm5@fedoraproject.org>