[release-1.33.2] Bump to v1.33.2
As the title says.

Rather than trying to insert the version bump to v1.33.2
into the main branch, I've created a branch that includes
the changes to Docker and Buildkit that should reduce
the noted bloat.  This does not include changes added for
upstream c/{project} bumps, XDG_RUNTIME changes, nor
FreeBSD.

This approach was used after discussions in my prior PR

This PR simiply bumps the version appropriately in the
branch so that we can create a release to vendor into
Podman v4.8.

[NO NEW TESTS NEEDED]

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>