build --all-platforms: skip some base "image" platforms
When figuring out which platforms the base images allow us to build
derived images for, screen out any images with non-empty artifactType
values.  Also screen out any which use empty values or the word
"unknown" in the OS and Architecture platform fields, and any
Architecture values that the compiler hasn't heard of.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>