[release-1.24] Cirrus: disable cross_build task
cross build task depends on `brew install`ing a go version that's
no longer supported upstream.

We will also be switching to TMT tests using centos stream, so cirrus
cross build task which will use the latest available version of golang
on centos stream. So disabling this task in cirrus shouldn't be too big
a deal.

It's only being commented out so we don't forget about it when we switch
to TMT.

Ref: https://cirrus-ci.com/build/6451305468657664

[NO NEW TESTS NEEDED]

Signed-off-by: Lokesh Mandvekar <lsm5@redhat.com>