Makefile: set GOTOOLCHAIN=local
Set GOTOOLCHAIN=local for all make targets, and for renovate.

Have the "vendor" target (and by extension, "vendor-in-container", which
our validation in CI uses) clear any "toolchain" directive that might
have been added to go.mod through manual invocations of the compiler.

At this point, we probably don't need to be checking for Go module
support, so switch to assuming it's available.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>