Execute cross-build task using PW pool
Previously this task ran using the Cirrus-CI compute service, consuming
compute credits.  However, since podman is already using a persistent
worker pool for CI, it can also be leveraged for use here.  Since Mac
resources are relatively expensive, it also makes financial sense to
re-use infrastructure where possible.

Note: This change also restricts the cross-build task to only running on
the default branch.  This is necessary because the PW Pool environment
is always rolling forward and may not match what was in place at the
time a release-branch was created.

Signed-off-by: Chris Evich <cevich@redhat.com>