tests: set _CONTAINERS_USERNS_CONFIGURED=done for libnetwork
Starting in common 0.59.1, github.com/containers/common's libnetwork no
longer attempts to detect when it's running in a user namespace to
decide where a lock file that it uses will be.  Since this test is doing
user namespace setup on its own as part of the test, we need to
explicitly tell it to not use the one that belongs to the node's root
user.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>