[release-1.33] Properly validate cache IDs and sources
The `--mount type=cache` argument to the `RUN` instruction in
Dockerfiles was using `filepath.Join` on user input, allowing
crafted paths to be used to gain access to paths on the host,
when the command should normally be limited only to Buildah;s own
cache and context directories. Switch to `filepath.SecureJoin` to
resolve the issue.

Fixes CVE-2024-9675
https://issues.redhat.com/browse/RHEL-61842

Signed-off-by: Matt Heon <mheon@redhat.com>
(cherry picked from commit 873beddbab81ad794255342568300ad904e54316)
Signed-off-by: tomsweeneyredhat <tsweeney@redhat.com>