pkg/overlay: cleanups
Change generateOverlayStructure() to not return its first argument
unchanged, since both of its callers already have that value, and adjust
a few error messages.

In the Linux MountWithOptions(), ensure that, if UpperDirOptionFragment
and WorkDirOptionFragment values were specified, they are absolute
paths, otherwise place them under the top-level parent of the various
directories we have for this mount.

Update a number of comments.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>