CI: remove dependencies on the cross-build task
Remove the bits of the CI configuration that prevent some jobs from
starting unless/until the cross-compile tests have passed.  That costs
more time than we're willing to wait, at this time at least.  Use the
skip directive to avoid having to get all of the dependencies and
only_if directives to agree.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>