Accept image names as sources for cache mounts
Allow the "from" option for cache mounts to refer to an image, which, if
we attempt to mount read-write, is mounted using an overlay to avoid
getting errors if the process we're running with the cache mount
attempts to write to it.

Add logic to clean up the bind mount in Builder.getBindMount() that's
triggered if Builder.runSetupVolumeMounts() returns an error.

When creating cache directories for cache mounts, process the
container's ID mappings.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>